Refactor chat_with_repo modules into packages
