+++ repo_agent/chat_with_repo/data_access/json_handler.py
@@ -1,3 +1,5 @@
+"""Utilities for loading and searching project JSON data."""
+
 import json
 import sys
