+++ repo_agent/chat_with_repo/rag.py
@@ -2,15 +2,15 @@ import json
 
 from llama_index.llms.openai import OpenAI
 
-from repo_agent.chat_with_repo.json_handler import JsonFileProcessor
+from repo_agent.chat_with_repo.data_access.json_handler import JsonFileProcessor
 from repo_agent.chat_with_repo.prompt import (
     query_generation_template,
     rag_ar_template,
     rag_template,
     relevance_ranking_chat_template,
 )
-from repo_agent.chat_with_repo.text_analysis_tool import TextAnalysisTool
-from repo_agent.chat_with_repo.vector_store_manager import VectorStoreManager
+from repo_agent.chat_with_repo.data_access.text_analysis_tool import TextAnalysisTool
+from repo_agent.chat_with_repo.vector_store.vector_store_manager import VectorStoreManager
 from repo_agent.log import logger
