+++ repo_agent/chat_with_repo/ui/__init__.py
@@ -1,0 +1,5 @@
+"""User interface components for interacting with the assistant."""
+
+from .gradio_interface import GradioInterface
+
+__all__ = ["GradioInterface"]
