+++ repo_agent/chat_with_repo/ui/gradio_interface.py
@@ -1,3 +1,5 @@
+"""Gradio-based user interface for interacting with the assistant."""
+
 import gradio as gr
 import markdown
