+++ repo_agent/chat_with_repo/vector_store/__init__.py
@@ -1,0 +1,5 @@
+"""Vector store management utilities for the chat_with_repo module."""
+
+from .vector_store_manager import VectorStoreManager
+
+__all__ = ["VectorStoreManager"]
